Remove int cast

+++ lib/Dialect/AIE/IR/AIEDialect.cpp
@@ -1212,11 +1212,11 @@ LogicalResult ShimDMAOp::verify() {
     }
   }
 
-  if ((int)inputChannels.size() >
+  if (inputChannels.size() >
       getTileOp().getNumSourceConnections(WireBundle::DMA))
     return emitOpError("uses more input channels than available on this tile");
 
-  if ((int)outputChannels.size() >
+  if (outputChannels.size() >
       getTileOp().getNumDestConnections(WireBundle::DMA))
     return emitOpError("uses more output channels than available on this tile");
 
@@ -1386,11 +1386,11 @@ LogicalResult MemOp::verify() {
                << "allocOp in MemOp region should have an id attribute";
   }
 
-  if ((int)inputChannels.size() >
+  if (inputChannels.size() >
       getTileOp().getNumSourceConnections(WireBundle::DMA))
     return emitOpError("uses more input channels than available on this tile");
 
-  if ((int)outputChannels.size() >
+  if (outputChannels.size() >
       getTileOp().getNumDestConnections(WireBundle::DMA))
     return emitOpError("uses more output channels than available on this tile");
 
@@ -1497,11 +1497,11 @@ LogicalResult MemTileDMAOp::verify() {
     }
   }
 
-  if ((int)inputChannels.size() >
+  if (inputChannels.size() >
       getTileOp().getNumSourceConnections(WireBundle::DMA))
     return emitOpError("uses more input channels than available on this tile");
 
-  if ((int)outputChannels.size() >
+  if (outputChannels.size() >
       getTileOp().getNumDestConnections(WireBundle::DMA))
     return emitOpError("uses more output channels than available on this tile");
